fixup! Move version logging to pretests
It doesn't belong in a subtest that may or may not be invoked.

Yeah, I know, this has gotten larger than I intended.
