+++ subtests/docker_cli/version/version.py
@@ -17,14 +17,12 @@ Prerequisites
 None
 """
 
-import os.path
-from autotest.client import utils
 from dockertest import subtest
 from dockertest.output import OutputGood
 from dockertest.output import DockerVersion
 from dockertest.output import mustpass
 from dockertest.dockercmd import DockerCmd
-from dockertest.docker_daemon import SocketClient, which_docker
+from dockertest.docker_daemon import SocketClient
 
 
 class version(subtest.Subtest):
@@ -43,20 +41,7 @@ class version(subtest.Subtest):
         # Raise exception on Go Panic or usage help message
         outputgood = OutputGood(self.stuff['cmdresult'])
         docker_version = DockerVersion(outputgood.stdout_strip)
-        info = ("docker version client: %s server %s"
-                % (docker_version.client, docker_version.server))
-        self.loginfo("Found %s", info)
-        with open(os.path.join(self.job.sysinfo.sysinfodir,
-                               'docker_version'), 'wb') as info_file:
-            info_file.write("%s\n" % info)
-        with open(os.path.join(self.job.sysinfo.sysinfodir,
-                               'docker_rpm'), 'wb') as rpm_file:
-            rpm_file.write(self._docker_rpm())
         self.verify_version(docker_version)
-
-    @staticmethod
-    def _docker_rpm():
-        return utils.run("rpm -q %s" % which_docker()).stdout
 
     def verify_version(self, docker_version):
         # TODO: Make URL to daemon configurable
